requirements.txt
gspread==3.6.0
Flask_Dance==3.0.0
Flask==1.1.2
Werkzeug==1.0.1
pandas==1.2.0
python-dotenv==0.15.0
